pyproject.toml
[project]
name = "deep-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "duckduckgo-search>=8.0.5",
    "langchain>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-google-genai>=2.1.6",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
